Implement Home Graph (beta); Add partial Analytics Graph logic

+++ app/src/main/java/com/example/gymify/core/util/HeightAndWeightFormatter.kt
@@ -8,7 +8,7 @@ fun formatWeight(weight: String, unit: UserWeightUnit): String {
     val numericWeight = weight.toFloatOrNull() ?: 0f
 
     return when(unit) {
-        UserWeightUnit.KG -> numericWeight.toString()
+        UserWeightUnit.KG -> numericWeight.toInt().toString()
         UserWeightUnit.LBS -> (numericWeight * 2.20462f).toInt().toString()
     }
 }
